% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_surveyctoapi.R
\name{check_scto_api}
\alias{check_scto_api}
\title{Check the SurveyCTO API has been authorized and stored as \code{auth}}
\usage{
check_scto_api(.auth = auth)
}
\arguments{
\item{.auth}{an authorization block from \code{{rsurveycto::scto_auth}}}
}
\value{
boolean
}
\description{
Check the SurveyCTO API has been authorized and stored as \code{auth}
}
\examples{
NULL
}
