% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_select_varlabs.R
\name{multisel_label}
\alias{multisel_label}
\title{Apply variable labels to multiselects}
\usage{
multisel_label(surveydata, thefields, thechoices)
}
\arguments{
\item{surveydata}{a tibble with survey data as processed by \code{scto_responses}}

\item{thefields}{a tibble with survey fields as processed by \code{read_questions}}

\item{thechoices}{a tibble with choices as processed by \code{read_choices}}
}
\value{
a tibble with variable labels for each column of a multiple_select question
}
\description{
Multiselects are complicated. They store responses to questions across multiple columns.
This function applies one label to each column, indicating both the question text (label) and response (label). These character strings are concatenated with a tilde \code{~} separator.
To do this, we process the labels first and then apply them to the columns of the response data.
At some point, I would like to have the ability to provide the user the choice whether these are dummy variables \verb{\{0,1\}}, missing values, \verb{\{NA, 1\}}, or booleans \verb{\{F, T\}}.
}
\examples{
NULL
}
