% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_choices.R
\name{read_choices}
\alias{read_choices}
\title{Extract response choices from a form definition}
\usage{
read_choices(theform)
}
\arguments{
\item{theform}{A survey form as imported by \code{scto_form}}
}
\value{
a tibble with response choices named as they appear on the XLS form \code{choices}
}
\description{
The form is available from the SurveyCTO.
It has three parts, including settings, fields, and choices.
We restructure this with dplyr and janitor so that it works nicely with other packages.
}
\examples{
NULL
}
