% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_questions.R
\name{read_questions}
\alias{read_questions}
\title{Read the questions from SurveyCTO fields}
\usage{
read_questions(theform)
}
\arguments{
\item{theform}{a SurveyCTO form as returned by \code{scto_form}}
}
\value{
a tibble of response lists, names, and values
}
\description{
This function reads the question list from a SurveyCTO data collection form.
The form is available from SurveyCTO. When downloaded from the API, it has four parts.
The question list is in the element titled "fields".
This function will declass the data.table so that it plays nicely with dplyr and tibble.
}
\examples{
NULL
}
