% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_getform.R
\name{scto_form}
\alias{scto_form}
\title{Get survey forms}
\usage{
scto_form(.formid, .auth = auth)
}
\arguments{
\item{.formid}{a character string that identifies the SurveyCTO form}

\item{.auth}{an authorization block generated by \code{rsurveycto::scto_auth()}}
}
\value{
A list object that corresponds to the sheets of survey form from SurveyCTO, which are typically downloaded as XLSX files. Class data.frame and data.table. The list has four elements: \code{settings}, \code{choices}, \code{fields}, and \code{formulasConvertedToStaticValues}.
}
\description{
Get the survey form definition for a specific survey form from SurveyCTO via API.  This form is a wrapper for \code{rsurveycto::scto_get_form_definitions()}.
}
\examples{
NULL
}
