% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_getdata.R
\name{scto_responses}
\alias{scto_responses}
\title{Get survey data}
\usage{
scto_responses(.formid = formid, .auth = auth)
}
\arguments{
\item{.formid}{a character string that identifies the SurveyCTO form}

\item{.auth}{an authorization block generated by \code{rsurveycto::scto_auth()}}
}
\value{
A \code{data.frame} of survey responses with every column in character format. As if it were loaded from CSV with \code{read.csv(..., colClasses="")}.
}
\description{
Get the survey responses for a specific survey form from SurveyCTO via API.  This form is a wrapper for \code{rsurveycto::scto_read()}.
}
\examples{
NULL

}
