% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_select_levels.R
\name{singsel_asfactor}
\alias{singsel_asfactor}
\title{Convert single select variables to factors}
\usage{
singsel_asfactor(surveydata, thefields, thechoices)
}
\arguments{
\item{surveydata}{a tibble with survey data as processed by \code{scto_responses}}

\item{thefields}{a tibble with survey fields as processed by \code{read_questions}}

\item{thechoices}{a tibble with choices as processed by \code{read_choices}}
}
\value{
a tibble with survey data where single select variables have variable labels in the \code{label} attribute
}
\description{
Convert single select variables to factors
}
\examples{
NULL
}
