% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_select_varlabs.R
\name{singsel_label}
\alias{singsel_label}
\title{Label single select questions}
\usage{
singsel_label(surveydata, thefields)
}
\arguments{
\item{surveydata}{a tibble with survey data as processed by \code{scto_responses}}

\item{thefields}{a tibble with survey fields as processed by \code{read_questions}}
}
\value{
a tibble with survey data where single select variables have variable labels in the \code{label} attribute
}
\description{
Label single select questions
}
\examples{
NULL
}
