% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven_disable_uniqueness_check.R
\name{validate_labelled}
\alias{validate_labelled}
\title{Title}
\usage{
validate_labelled(x)
}
\arguments{
\item{x}{an object with a labels attribute}
}
\value{
the same object
}
\description{
Title
}
\examples{
NULL
}
